put version checking unittests together
